Make type IDs cheap in caffe2
right now getting a type ID requires (1) a call to an external function (2) a branch to check the static initialization of the type ID.

This diff makes type type ID a constexpr. The only issue here is that it seems to be issuing a warning of the form:

```
caffe2/caffe2/core/typeid.h:215:57: warning: instantiation of variable 'caffe2::TypeIdBit<caffe2::Tensor<caffe2::CUDAContext> >::bit' required here, but no definition is available [-Wundefined-var-template]
    return reinterpret_cast<CaffeTypeId>(&TypeIdBit<T>::bit);
```

I'm not sure how to best get rid of this.